requirements.txt
fastapi~=0.115.12
pydantic~=2.11.2
python-dotenv~=1.1.0
protobuf~=5.29.4